refactor(components): reorganize component files and update imports
