+++ src/components/menu.tsx
@@ -2,13 +2,11 @@
 
 import { linksMenu } from "@/constants/linksMenu";
 import { motion } from "motion/react";
-import { useTheme } from "next-themes";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { ThemeSwitchMenu } from "../provider/switch-theme";
+import { ThemeSwitchMenu } from "./provider/switch-theme";
 
 export default function Menu() {
-  const { theme } = useTheme();
   const pathname = usePathname();
 
   return (
@@ -21,14 +19,6 @@ export default function Menu() {
       <div className="backdrop-blur-sm border-1 dark:border-white/30 flex h-12 items-center justify-between max-sm:w-sm  max-md:px-1 max-md:w-lg px-4  w-3xl  rounded-xl">
         <div>
           <Link className="flex justify-center font-sans items-center" href="/">
-            {/* <Image
-              suppressHydrationWarning
-              className="dark:text-white"
-              src={theme === "dark" ? "/ofs-light.svg" : "/ofs-dark.svg"}
-              width={36}
-              height={36}
-              alt="logo"
-            /> */}
             <span className="font-semi-bold max-sm:hidden text-3xl font-serif">
               OFS
             </span>
